move state/old/index into /state/index
